fix(api): use int64 for ASN fields to fix invalid OpenAPI v3 schema
Change LocalASN/PeerASN from uint32 to int64. uint32 maps to format:int32
in OpenAPI v3 (signed range only), making maximum:4294967295 invalid and
causing Kubernetes to reject all BGPRouter/BGPPeer creations.

int64 produces format:int64 in the CRD, supporting values up to
9223372036854775807 — well above the 4294967295 max ASN.

Add regression tests for large 4-byte ASNs (>2^31-1) in both BGPRouter
and BGPPeer, plus missing BGPPeer DeepCopy/JSON tests.

+++ api/bgp/v1alpha1/peer_types.go
@@ -43,7 +43,7 @@ type BGPPeerSpec struct {
 	// +kubebuilder:validation:Required
 	// +kubebuilder:validation:Minimum=1
 	// +kubebuilder:validation:Maximum=4294967295
-	PeerASN uint32 `json:"peerASN"`
+	PeerASN int64 `json:"peerASN"`
 
 	// Address is the remote peer's IPv4 or IPv6 address.
 	// +kubebuilder:validation:Required
